Implement javadocs for CFR
